refactor models
