+++ src/models/AbstractFactory/IShapeFactory.ts
@@ -1,0 +1,7 @@
+import { ICircle } from '../AbstractProduct/ICircle';
+import { IRectangle } from '../AbstractProduct/IRectangle';
+
+export interface IShapeFactory {
+    createCircle(): ICircle;
+    createRectangle(): IRectangle;
+}
