+++ src/models/AbstractProduct/ICircle.ts
