+++ src/models/AbstractProduct/IRectangle.ts
