+++ src/models/AbstractProduct/IShape.ts
