+++ src/models/ConcreteProduct/RectangleFill.ts
@@ -1,8 +1,8 @@
-import { CanvasSingleton } from '../CanvasSingleton';
-import { IRectangle } from '../IRectangle';
+import { IRectangle } from '../AbstractProduct/IRectangle';
 import { Point } from '../Point';
+import { CanvasSingleton } from '../CanvasSingleton';
 
-export class RectangleStroke implements IRectangle {
+export class RectangleFill implements IRectangle {
     constructor(
         private _startPosition: Point = new Point(0, 0),
         private _width: number = 0,
@@ -31,7 +31,7 @@ export class RectangleStroke implements IRectangle {
         CanvasSingleton.getInstance().canvas.addEventListener('mouseup', this._handleMouseUp);
     }
 
-    public stopPaint(): void {
+    stopPaint(): void {
         CanvasSingleton.getInstance().canvas.removeEventListener(
             'mousedown',
             this._handleMouseDown
@@ -44,18 +44,22 @@ export class RectangleStroke implements IRectangle {
     }
 
     private _handleMouseDown(e: MouseEvent): void {
-        CanvasSingleton.getInstance().isPainting = true;
         this._startPosition = new Point(e.offsetX, e.offsetY);
+        CanvasSingleton.getInstance().isPainting = true;
         CanvasSingleton.getInstance().context.beginPath();
     }
 
-    private _handleMouseMove(e: MouseEvent): void {}
+    private _handleMouseMove(e: MouseEvent): void {
+        if (CanvasSingleton.getInstance().isPainting) {
+        }
+    }
 
     private _handleMouseUp(e: MouseEvent): void {
         this._width = e.offsetX - this._startPosition.getX();
         this._height = e.offsetY - this._startPosition.getY();
+
         CanvasSingleton.getInstance().isPainting = false;
-        CanvasSingleton.getInstance().context.strokeRect(
+        CanvasSingleton.getInstance().context.fillRect(
             this._startPosition.getX(),
             this._startPosition.getY(),
             this._width,
