--- src/models/IShapeFactory.ts
@@ -1,7 +1,0 @@
-import { ICircle } from './ICircle';
-import { IRectangle } from './IRectangle';
-
-export interface IShapeFactory {
-    createCircle(): ICircle;
-    createRectangle(): IRectangle;
-}
